refactor template
